requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
bcrypt==4.0.1
python-multipart>=0.0.6
httpx>=0.25.0
pytest>=7.4.0